backend/requirements.txt
fastapi
uvicorn[standard]
openai
numpy
python-dotenv
pandas # Still useful for data manipulation if needed